added search_for_user function
when function is called user is given choices on how they would like to search for a user to send a request to (by last name, university, or major). It then returns the username of the user they would like to request if it matches a result in the database.